lib/src: Revert scaling in FOC device
Foc driver supports scaling internally
can receive negative and fractions.

+++ lib/src/foc.h
@@ -26,13 +26,13 @@ public:
 	int setGain(GainType gainController, double kp, double ki);
 	GainData getGain(GainType gainController);
 	int startFoc();
-	int setAngleOffset(int angleSh);
+	int setAngleOffset(double angleSh);
 	int setFixedSpeed(int fixedSpeed);
 	int setVfParam(double vq, double vd, int fixedSpeed);
 	int stopMotor();
 	double getTorqueSetValue();
 	int setOperationMode(MotorOpMode mode);
-	int getSpeedSetValue();
+	double getSpeedSetValue();
 	FocData getChanData();
 	~Foc();
 
@@ -40,15 +40,15 @@ private:
 	IIO_Driver *mFoc_IIO_Handle;
 	static const std::string kFocDriverName;
 
-	int mTargetSpeed;
+	double mTargetSpeed;
 	int mSpeedRRate;
 	bool mDoSpeedRamp;
 	std::thread mSpeedThread;
 	std::mutex mSpeedMutex;
 	void rampSpeed(void);
 
-	int mTargetTorque;
-	int mTorRRate;
+	double mTargetTorque;
+	double mTorRRate;
 	bool mDoTorRamp;
 	std::thread mTorThread;
 	std::mutex mTorMutex;
